.pipelines/prchecks/CveSpecFilePRCheck/azure-function/requirements.txt
azure-functions>=1.18.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
PyJWT>=2.8.0
requests>=2.31.0
cryptography>=41.0.0
